token issue solved

+++ app/Http/Controllers/Country/DeleteCountryController.php
@@ -10,12 +10,18 @@ class DeleteCountryController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('auth');
-        $this->middleware('admin');
+        $this->middleware('auth:sanctum');
     }
 
-    public function destroy($id)
+    public function destroy(Request $request,$id)
     {
+
+        $user=$request->user();
+        
+        if ($user->role->name !== 'admin') {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         $country = Country::findOrFail($id);
         $country->delete();
